+++ src/component/Pages/Fiji/Fiji.jsx
@@ -12,9 +12,19 @@ import shareFiji from '../../../images/fiji/history/shareFiji.mp4';
             <div>
 				<section className="internal-page-wrapper">
 					<div className="container"> 
-                    <video width="100%" controls >
-                        <source src={shareFiji} type="video/mp4"/>
-                    </video>
+						<div className="row justify-content-center text-center">
+							<div className="col-12">
+								<div className="title-wraper">
+									Fiji - Home Automation
+								</div>
+								<div className="brief-description">
+									What if your home helped you? 
+								</div>
+							</div> 
+						</div>
+                        <video width="100%" controls >
+                            <source src={shareFiji} type="video/mp4"/>
+                        </video>
 					</div>
 				</section>
 			</div>
